Source File
This is a demo of interfacing a multi coin acceptor usually used in
pisonet timers here in the Philippines. The coin acceptor outputs pulses
depending on the coin denomination. The coin pin of the acceptor is
connected to pin D3 of the Gizduino Board which is an external interrupt
pin of the AVR Microcontroller.